requirements.txt
pygame==2.6.0
pygame-ce==2.5.1
pygame-menu==4.4.3
pygame-textinput==1.0.1
python-rtmidi==1.5.8
